+++ R/data.R
@@ -1,0 +1,35 @@
+#' GIS advisory dataset for Hurricane Sandy Adv 18
+#' @source \url{http://www.nhc.noaa.gov/gis/archive_forecast_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
+"adv"
+
+#' Atlantic cyclones for 2012
+#' @source \url{http://www.nhc.noaa.gov/archive/2012/}
+"al.2012"
+
+#' Strike probabilities for Hurricane Katrina (AL122005)
+#' @source \url{http://www.nhc.noaa.gov/archive/2005/KATRINA.shtml?}
+"al122005_prblty"
+
+#' Wind speed probabilities for Hurricane Sandy (AL182012)
+#' @source \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
+"al182012_wndprb"
+
+#' Forecast/Advisory and Wind Speed Probabilities for Hurricane Sandy (AL182012)
+#' @source \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
+"al182012"
+
+#' Forecast/Advisory for Hurricane Sandy (AL182012)
+#' @source \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
+"al182012_fstadv"
+
+#' GIS storm surge shapefile dataset for Hurricane Sandy (AL182012)
+#' @source \url{http://www.nhc.noaa.gov/gis/archive_psurge_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
+"ss"
+
+#' GIS windfield and forecast wind radii for Hurricane Sandy (AL182012)
+#' @source \url{http://www.nhc.noaa.gov/gis/archive_forecast_info_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
+"wf"
+
+#' GIS wind speed probabilities for Hurricane Sandy (AL182012)
+#' @source \url{http://www.nhc.noaa.gov/gis/archive_wsp.php}
+"wsp"