+++ tests/testthat/test_get_storm_data.R
@@ -1,21 +1,15 @@
 context("Get Storm Data")
 
-# Set timeout options
-opt.timeout <- getOption("rrricanes.http_timeout")
-opt.attempts <- getOption("rrricanes.http_attempts")
-options("rrricanes.http_timeout" = 1)
-options("rrricanes.http_attempts" = 5)
-
-## ---- Base Data --------------------------------------------------------------
-## ---- * 2008, AL -------------------------------------------------------------
-df.al_2017_storm_data <- get_storms(year = 2017, basin = "AL") %>%
-    dplyr::slice(1) %>%
-    .$Link %>%
-    get_storm_data(products = c("discus", "fstadv"))
 load(system.file("extdata", "al_2017_storm_data.Rda", package = "rrricanes"))
 
 ## ---- get_storm_data() -------------------------------------------------------
 test_that("get_storm_data()", {
+    skip_on_cran()
+    df.al_2017_storm_data <- get_storms(year = 2017, basin = "AL") %>%
+        dplyr::slice(1) %>%
+        .$Link %>%
+        get_storm_data(products = c("discus", "fstadv"))
+
     ## ---- * 2017, AL, 01 -----------------------------------------------------
     expect_identical(al_2017_storm_data, df.al_2017_storm_data)
     ## ---- * Errors -----------------------------------------------------------
@@ -26,7 +20,3 @@ test_that("get_storm_data()", {
     expect_error(get_storm_data(),
                  "argument \"link\" is missing, with no default")
 })
-
-# Reset options
-options("rrricanes.http_timeout" = opt.timeout)
-options("rrricanes.http_attempts" = opt.attempts)