+++ tests/testthat/test_posest.R
@@ -1,17 +1,5 @@
 context("Position Estimates (posest)")
 
-# Set timeout options
-opt.timeout <- getOption("rrricanes.http_timeout")
-opt.attempts <- getOption("rrricanes.http_attempts")
-options("rrricanes.http_timeout" = 1)
-options("rrricanes.http_attempts" = 5)
-
-## ---- 2008, AL ---------------------------------------------------------------
-al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-
-## ---- Base Data --------------------------------------------------------------
-## ---- * 2008, AL -------------------------------------------------------------
-df.al092008.posest <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_posest()
 load(system.file("extdata", "al092008.posest.Rda", package = "rrricanes"))
 
 ## ---- Dataframe Skeleton -----------------------------------------------------
@@ -29,9 +17,8 @@ test_that("Dataframe Skeleton", {
 
 ## ---- Test get_posest() ------------------------------------------------------
 test_that("Test get_posest()", {
+    skip_on_cran()
+    al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+    df.al092008.posest <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_posest()
     expect_identical(al092008.posest, df.al092008.posest)
 })
-
-# Reset options
-options("rrricanes.http_timeout" = opt.timeout)
-options("rrricanes.http_attempts" = opt.attempts)