+++ tests/testthat/test_public.R
@@ -1,24 +1,11 @@
 context("Public Advisory (public)")
 
-# Set timeout options
-opt.timeout <- getOption("rrricanes.http_timeout")
-opt.attempts <- getOption("rrricanes.http_attempts")
-options("rrricanes.http_timeout" = 1)
-options("rrricanes.http_attempts" = 5)
-
-## ---- 2008, AL ---------------------------------------------------------------
-al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-
-## ---- Base Data --------------------------------------------------------------
-## ---- * 2008, AL -------------------------------------------------------------
-df.al092008.public <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_public()
 load(system.file("extdata", "al092008.public.Rda", package = "rrricanes"))
 
 ## ---- Test get_public() ------------------------------------------------------
 test_that("Test get_public()", {
-    expect_equal(al092008.public, df.al092008.public)
+    skip_on_cran()
+    al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+    df.al092008.public <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_public()
+    testthat::expect_equal(al092008.public, df.al092008.public)
 })
-
-# Reset options
-options("rrricanes.http_timeout" = opt.timeout)
-options("rrricanes.http_attempts" = opt.attempts)